restructure into PackageDeprecationManagementService
