+++ tests/NuGetGallery.Facts/Controllers/ManageDeprecationJsonApiControllerFacts.cs
@@ -6,9 +6,8 @@ using System.Linq;
 using System.Net;
 using System.Threading.Tasks;
 using System.Web.Mvc;
+using Moq;
 using Newtonsoft.Json.Linq;
-using NuGet.Services.Entities;
-using NuGet.Versioning;
 using NuGetGallery.Framework;
 using NuGetGallery.RequestModels;
 using Xunit;
@@ -20,13 +19,15 @@ namespace NuGetGallery.Controllers
         public class TheGetAlternatePackageVersionsMethod : TestContainer
         {
             [Fact]
-            public void ReturnsEmptyIfNoVersions()
+            public void ReturnsProperJsonResult()
             {
                 // Arrange
                 var id = "Crested.Gecko";
-                GetMock<IPackageService>()
-                    .Setup(x => x.FindPackagesById(id, PackageDeprecationFieldsToInclude.None))
-                    .Returns(new Package[0]);
+
+                var versions = new[] { "1.0.0", "2.0.0" };
+                GetMock<IPackageDeprecationManagementService>()
+                    .Setup(x => x.GetPossibleAlternatePackageVersions(id))
+                    .Returns(versions);
 
                 var controller = GetController<ManageDeprecationJsonApiController>();
 
@@ -36,702 +37,78 @@ namespace NuGetGallery.Controllers
                 // Assert
                 Assert.Equal(JsonRequestBehavior.AllowGet, result.JsonRequestBehavior);
                 Assert.Equal((int)HttpStatusCode.OK, controller.Response.StatusCode);
-                Assert.Empty((IEnumerable<string>)result.Data);
-            }
-
-            [Fact]
-            public void ReturnsAllAvailableVersionsInReverseVersionOrder()
-            {
-                // Arrange
-                var id = "Crested.Gecko";
-                var firstPackage = new Package
-                {
-                    Version = "1.0.0+build",
-                    Listed = true
-                };
-
-                var secondPackage = new Package
-                {
-                    Version = "2.0.0+build",
-                    Listed = true
-                };
-
-                var thirdPackage = new Package
-                {
-                    Version = "3.0.0+build",
-                    Listed = true
-                };
-
-                var deletedPackage = new Package
-                {
-                    Version = "2.1.0-deleted",
-                    PackageStatusKey = PackageStatus.Deleted,
-                    Listed = true
-                };
-
-                var validatingPackage = new Package
-                {
-                    Version = "2.2.0-validating",
-                    PackageStatusKey = PackageStatus.Validating,
-                    Listed = true
-                };
-
-                var failedValidationPackage = new Package
-                {
-                    Version = "2.3.0-failedValidation",
-                    PackageStatusKey = PackageStatus.FailedValidation,
-                    Listed = true
-                };
-
-                var unlistedPackage = new Package
-                {
-                    Version = "2.4.0-unlisted",
-                    Listed = false
-                };
-
-                var packages = new[]
-                {
-                    firstPackage,
-                    deletedPackage,
-                    validatingPackage,
-                    failedValidationPackage,
-                    thirdPackage,
-                    secondPackage,
-                    unlistedPackage
-                };
-
-                GetMock<IPackageService>()
-                    .Setup(x => x.FindPackagesById(id, PackageDeprecationFieldsToInclude.None))
-                    .Returns(packages);
-
-                var controller = GetController<ManageDeprecationJsonApiController>();
-
-                // Act
-                var result = controller.GetAlternatePackageVersions(id);
-
-                // Assert
-                Assert.Equal(JsonRequestBehavior.AllowGet, result.JsonRequestBehavior);
-                Assert.Equal((int)HttpStatusCode.OK, controller.Response.StatusCode);
-
-                var expected = new[] { thirdPackage, secondPackage, firstPackage }
-                    .Select(v => NuGetVersion.Parse(v.Version).ToNormalizedString())
-                    .ToList();
-
-                Assert.Equal(expected, result.Data);
+                Assert.Equal(versions, result.Data);
             }
         }
 
         public class TheDeprecateMethod : TestContainer
         {
-            [Fact]
-            public async Task ReturnsBadRequestIfOtherAndNoCustomMessage()
-            {
-                // Arrange
-                var controller = GetController<ManageDeprecationJsonApiController>();
-
-                // Act
-                var result = await controller.Deprecate(
-                    CreateDeprecatePackageRequest("id", isOther: true));
-
-                // Assert
-                AssertErrorResponse(controller, result, HttpStatusCode.BadRequest, Strings.DeprecatePackage_CustomMessageRequired);
-            }
-
-            public static IEnumerable<object[]> ReturnsBadRequestIfNoVersions_Data =
-                MemberDataHelper.AsDataSet(null, new string[0]);
+            public static IEnumerable<object[]> ReturnsProperJsonResult_Data =
+                MemberDataHelper.Combine(
+                    Enumerable
+                        .Repeat(
+                            MemberDataHelper.BooleanDataSet(), 4)
+                        .ToArray());
 
             [Theory]
-            [MemberData(nameof(ReturnsBadRequestIfNoVersions_Data))]
-            public async Task ReturnsBadRequestIfNoVersions(IEnumerable<string> versions)
+            [MemberData(nameof(ReturnsProperJsonResult_Data))]
+            public async Task ReturnsProperJsonResult(
+                bool isLegacy,
+                bool hasCriticalBugs,
+                bool isOther,
+                bool success)
             {
                 // Arrange
-                var currentUser = TestUtility.FakeUser;
+                var id = "Crested.Gecko";
+                var versions = new[] { "1.0.0", "2.0.0" };
+                var alternateId = "alt.Id";
+                var alternateVersion = "3.0.0";
+                var customMessage = "custom";
+
+                var currentUser = Get<Fakes>().User;
+
+                var errorStatus = HttpStatusCode.InternalServerError;
+                var errorMessage = "woops";
+                GetMock<IPackageDeprecationManagementService>()
+                    .Setup(x => x.UpdateDeprecation(
+                        currentUser,
+                        id,
+                        versions,
+                        isLegacy,
+                        hasCriticalBugs,
+                        isOther,
+                        alternateId,
+                        alternateVersion,
+                        customMessage))
+                    .ReturnsAsync(success ? null : new UpdateDeprecationError(errorStatus, errorMessage));
 
                 var controller = GetController<ManageDeprecationJsonApiController>();
                 controller.SetCurrentUser(currentUser);
 
-                // Act
-                var result = await controller.Deprecate(
-                    CreateDeprecatePackageRequest("id", versions));
-
-                // Assert
-                AssertErrorResponse(controller, result, HttpStatusCode.BadRequest, Strings.DeprecatePackage_NoVersions);
-            }
-
-            [Fact]
-            public async Task ReturnsBadRequestIfLongCustomMessage()
-            {
-                // Arrange
-                var currentUser = TestUtility.FakeUser;
-                var controller = GetController<ManageDeprecationJsonApiController>();
-                controller.SetCurrentUser(currentUser);
-
-                var customMessage = new string('a', 1001);
-
-                // Act
-                var result = await controller.Deprecate(
-                    CreateDeprecatePackageRequest(
-                        "id",
-                        new[] { "1.0.0" },
-                        customMessage: customMessage));
-
-                // Assert
-                AssertErrorResponse(
-                    controller,
-                    result,
-                    HttpStatusCode.BadRequest,
-                    string.Format(Strings.DeprecatePackage_CustomMessageTooLong, 1000));
-            }
-
-            public static IEnumerable<object[]> ReturnsNotFoundIfNoPackagesOrRegistrationMissing_Data
-            {
-                get
-                {
-                    var packageWithNullRegistration = new Package();
-                    return MemberDataHelper.AsDataSet(
-                        new Package[0],
-                        new[] { packageWithNullRegistration });
-                }
-            }
-
-            [Theory]
-            [MemberData(nameof(ReturnsNotFoundIfNoPackagesOrRegistrationMissing_Data))]
-            public async Task ReturnsNotFoundIfNoPackagesOrRegistrationMissing(IEnumerable<Package> packages)
-            {
-                // Arrange
-                var id = "id";
-                var currentUser = TestUtility.FakeUser;
-
-                var packageService = GetMock<IPackageService>();
-                packageService
-                    .Setup(x => x.FindPackagesById(id, PackageDeprecationFieldsToInclude.DeprecationAndRelationships))
-                    .Returns(packages.ToList())
-                    .Verifiable();
-
-                var controller = GetController<ManageDeprecationJsonApiController>();
-                controller.SetCurrentUser(currentUser);
-
-                // Act
-                var result = await controller.Deprecate(
-                    CreateDeprecatePackageRequest("id", new[] { "1.0.0" }));
-
-                // Assert
-                AssertErrorResponse(controller, result, HttpStatusCode.NotFound, string.Format(Strings.DeprecatePackage_MissingRegistration, id));
-                packageService.Verify();
-            }
-
-
-            [Fact]
-            public async Task ReturnsForbiddenIfFeatureFlagDisabled()
-            {
-                // Arrange
-                var id = "id";
-                var currentUser = TestUtility.FakeUser;
-
-                var registration = new PackageRegistration
-                {
-                    Id = id
-                };
-
-                var package = new Package
-                {
-                    PackageRegistration = registration
-                };
-
-                var packageService = GetMock<IPackageService>();
-                packageService
-                    .Setup(x => x.FindPackagesById(id, PackageDeprecationFieldsToInclude.DeprecationAndRelationships))
-                    .Returns(new[] { package })
-                    .Verifiable();
-
-                var featureFlagService = GetMock<IFeatureFlagService>();
-                featureFlagService
-                    .Setup(x => x.IsManageDeprecationEnabled(currentUser, registration))
-                    .Returns(false)
-                    .Verifiable();
-
-                var controller = GetController<ManageDeprecationJsonApiController>();
-                controller.SetCurrentUser(currentUser);
-
-                // Act
-                var result = await controller.Deprecate(
-                    CreateDeprecatePackageRequest(id, new[] { "1.0.0" }));
-
-                // Assert
-                AssertErrorResponse(controller, result, HttpStatusCode.Forbidden, Strings.DeprecatePackage_Forbidden);
-                featureFlagService.Verify();
-            }
-
-            public static IEnumerable<object[]> NotOwner_Data
-            {
-                get
-                {
-                    yield return new object[]
-                    {
-                        null,
-                        TestUtility.FakeUser
-                    };
-
-                    yield return new object[]
-                    {
-                        TestUtility.FakeUser,
-                        new User { Key = 5535 }
-                    };
-                }
-            }
-
-            [Theory]
-            [MemberData(nameof(NotOwner_Data))]
-            public async Task ReturnsForbiddenIfNotOwner(User currentUser, User owner)
-            {
-                // Arrange
-                var id = "id";
-
-                var registration = new PackageRegistration
-                {
-                    Id = id
-                };
-
-                var featureFlagService = GetMock<IFeatureFlagService>();
-                featureFlagService
-                    .Setup(x => x.IsManageDeprecationEnabled(currentUser, registration))
-                    .Returns(true)
-                    .Verifiable();
-
-                registration.Owners.Add(owner);
-
-                var package = new Package
-                {
-                    PackageRegistration = registration
-                };
-
-                var packageService = GetMock<IPackageService>();
-                packageService
-                    .Setup(x => x.FindPackagesById(id, PackageDeprecationFieldsToInclude.DeprecationAndRelationships))
-                    .Returns(new[] { package })
-                    .Verifiable();
-
-                var controller = GetController<ManageDeprecationJsonApiController>();
-                controller.SetCurrentUser(currentUser);
-
-                // Act
-                var result = await controller.Deprecate(
-                    CreateDeprecatePackageRequest(id, new[] { "1.0.0" }));
-
-                // Assert
-                AssertErrorResponse(controller, result, HttpStatusCode.Forbidden, Strings.DeprecatePackage_Forbidden);
-                featureFlagService.Verify();
-                packageService.Verify();
-            }
-
-            public static IEnumerable<object[]> Owner_Data
-            {
-                get
-                {
-                    yield return new object[]
-                    {
-                        TestUtility.FakeUser,
-                        TestUtility.FakeUser
-                    };
-
-                    yield return new object[]
-                    {
-                        TestUtility.FakeAdminUser,
-                        TestUtility.FakeUser
-                    };
-
-                    yield return new object[]
-                    {
-                        TestUtility.FakeOrganizationAdmin,
-                        TestUtility.FakeOrganization
-                    };
-
-                    yield return new object[]
-                    {
-                        TestUtility.FakeOrganizationCollaborator,
-                        TestUtility.FakeOrganization
-                    };
-                }
-            }
-
-            [Theory]
-            [MemberData(nameof(Owner_Data))]
-            public async Task ReturnsForbiddenIfLocked(User currentUser, User owner)
-            {
-                // Arrange
-                var id = "id";
-
-                var registration = new PackageRegistration
+                var request = new DeprecatePackageRequest
                 {
                     Id = id,
-                    IsLocked = true
+                    Versions = versions,
+                    IsLegacy = isLegacy,
+                    HasCriticalBugs = hasCriticalBugs,
+                    IsOther = isOther,
+                    AlternatePackageId = alternateId,
+                    AlternatePackageVersion = alternateVersion,
+                    CustomMessage = customMessage
                 };
-
-                var featureFlagService = GetMock<IFeatureFlagService>();
-                featureFlagService
-                    .Setup(x => x.IsManageDeprecationEnabled(currentUser, registration))
-                    .Returns(true)
-                    .Verifiable();
-
-                registration.Owners.Add(owner);
-
-                var package = new Package
-                {
-                    PackageRegistration = registration
-                };
-
-                var packageService = GetMock<IPackageService>();
-                packageService
-                    .Setup(x => x.FindPackagesById(id, PackageDeprecationFieldsToInclude.DeprecationAndRelationships))
-                    .Returns(new[] { package })
-                    .Verifiable();
-
-                var controller = GetController<ManageDeprecationJsonApiController>();
-                controller.SetCurrentUser(currentUser);
 
                 // Act
-                var result = await controller.Deprecate(
-                    CreateDeprecatePackageRequest(id, new[] { "1.0.0" }));
+                var result = await controller.Deprecate(request);
 
                 // Assert
-                AssertErrorResponse(
-                    controller,
-                    result,
-                    HttpStatusCode.Forbidden,
-                    string.Format(Strings.DeprecatePackage_Locked, id));
-                featureFlagService.Verify();
-                packageService.Verify();
-            }
-
-            [Theory]
-            [MemberData(nameof(Owner_Data))]
-            public async Task ReturnsNotFoundIfAlternatePackageRegistrationMissing(User currentUser, User owner)
-            {
-                // Arrange
-                var id = "id";
-
-                var registration = new PackageRegistration
+                if (success)
                 {
-                    Id = id
-                };
-
-                var featureFlagService = GetMock<IFeatureFlagService>();
-                featureFlagService
-                    .Setup(x => x.IsManageDeprecationEnabled(currentUser, registration))
-                    .Returns(true)
-                    .Verifiable();
-
-                registration.Owners.Add(owner);
-
-                var package = new Package
+                    AssertSuccessResponse(controller);
+                }
+                else
                 {
-                    PackageRegistration = registration
-                };
-
-                var packageService = GetMock<IPackageService>();
-                packageService
-                    .Setup(x => x.FindPackagesById(id, PackageDeprecationFieldsToInclude.DeprecationAndRelationships))
-                    .Returns(new[] { package })
-                    .Verifiable();
-
-                var alternatePackageId = "altId";
-                packageService
-                    .Setup(x => x.FindPackageRegistrationById(alternatePackageId))
-                    .Returns((PackageRegistration)null)
-                    .Verifiable();
-
-                var controller = GetController<ManageDeprecationJsonApiController>();
-                controller.SetCurrentUser(currentUser);
-
-                // Act
-                var result = await controller.Deprecate(
-                    CreateDeprecatePackageRequest(
-                        id, 
-                        new[] { "1.0.0" }, 
-                        alternatePackageId: alternatePackageId));
-
-                // Assert
-                AssertErrorResponse(
-                    controller,
-                    result,
-                    HttpStatusCode.NotFound,
-                    string.Format(Strings.DeprecatePackage_NoAlternatePackageRegistration, alternatePackageId));
-                featureFlagService.Verify();
-                packageService.Verify();
-            }
-
-            [Theory]
-            [MemberData(nameof(Owner_Data))]
-            public async Task ReturnsNotFoundIfAlternatePackageVersionMissing(User currentUser, User owner)
-            {
-                // Arrange
-                var id = "id";
-
-                var registration = new PackageRegistration
-                {
-                    Id = id
-                };
-
-                var featureFlagService = GetMock<IFeatureFlagService>();
-                featureFlagService
-                    .Setup(x => x.IsManageDeprecationEnabled(currentUser, registration))
-                    .Returns(true)
-                    .Verifiable();
-
-                registration.Owners.Add(owner);
-
-                var package = new Package
-                {
-                    PackageRegistration = registration
-                };
-
-                var packageService = GetMock<IPackageService>();
-                packageService
-                    .Setup(x => x.FindPackagesById(id, PackageDeprecationFieldsToInclude.DeprecationAndRelationships))
-                    .Returns(new[] { package })
-                    .Verifiable();
-
-                var alternatePackageId = "altId";
-                var alternatePackageVersion = "1.2.3-alt";
-                packageService
-                    .Setup(x => x.FindPackageByIdAndVersionStrict(alternatePackageId, alternatePackageVersion))
-                    .Returns((Package)null)
-                    .Verifiable();
-
-                var controller = GetController<ManageDeprecationJsonApiController>();
-                controller.SetCurrentUser(currentUser);
-
-                // Act
-                var result = await controller.Deprecate(
-                    CreateDeprecatePackageRequest(
-                        id, 
-                        new[] { "1.0.0" }, 
-                        alternatePackageId: alternatePackageId, 
-                        alternatePackageVersion: alternatePackageVersion));
-
-                // Assert
-                AssertErrorResponse(
-                    controller,
-                    result,
-                    HttpStatusCode.NotFound,
-                    string.Format(Strings.DeprecatePackage_NoAlternatePackage, alternatePackageId, alternatePackageVersion));
-                featureFlagService.Verify();
-                packageService.Verify();
-            }
-
-            [Theory]
-            [MemberData(nameof(Owner_Data))]
-            public async Task ReturnsNotFoundIfVersionMissing(User currentUser, User owner)
-            {
-                // Arrange
-                var id = "id";
-
-                var registration = new PackageRegistration
-                {
-                    Id = id
-                };
-
-                var featureFlagService = GetMock<IFeatureFlagService>();
-                featureFlagService
-                    .Setup(x => x.IsManageDeprecationEnabled(currentUser, registration))
-                    .Returns(true)
-                    .Verifiable();
-
-                registration.Owners.Add(owner);
-
-                var package = new Package
-                {
-                    NormalizedVersion = "2.3.4",
-                    PackageRegistration = registration
-                };
-
-                var packageService = GetMock<IPackageService>();
-                packageService
-                    .Setup(x => x.FindPackagesById(id, PackageDeprecationFieldsToInclude.DeprecationAndRelationships))
-                    .Returns(new[] { package })
-                    .Verifiable();
-
-                var controller = GetController<ManageDeprecationJsonApiController>();
-                controller.SetCurrentUser(currentUser);
-
-                // Act
-                var result = await controller.Deprecate(
-                    CreateDeprecatePackageRequest(id, new[] { "1.0.0" }));
-
-                // Assert
-                AssertErrorResponse(
-                    controller,
-                    result,
-                    HttpStatusCode.NotFound,
-                    string.Format(Strings.DeprecatePackage_MissingVersion, id));
-                featureFlagService.Verify();
-                packageService.Verify();
-            }
-
-            [Theory]
-            [MemberData(nameof(Owner_Data))]
-            public async Task ReturnsNotFoundIfSomeVersionMissing(User currentUser, User owner)
-            {
-                // Arrange
-                var id = "id";
-
-                var registration = new PackageRegistration
-                {
-                    Id = id
-                };
-
-                var featureFlagService = GetMock<IFeatureFlagService>();
-                featureFlagService
-                    .Setup(x => x.IsManageDeprecationEnabled(currentUser, registration))
-                    .Returns(true)
-                    .Verifiable();
-
-                registration.Owners.Add(owner);
-
-                var package = new Package
-                {
-                    NormalizedVersion = "2.3.4",
-                    PackageRegistration = registration
-                };
-
-                var packageService = GetMock<IPackageService>();
-                packageService
-                    .Setup(x => x.FindPackagesById(id, PackageDeprecationFieldsToInclude.DeprecationAndRelationships))
-                    .Returns(new[] { package })
-                    .Verifiable();
-
-                var controller = GetController<ManageDeprecationJsonApiController>();
-                controller.SetCurrentUser(currentUser);
-
-                // Act
-                var result = await controller.Deprecate(
-                    CreateDeprecatePackageRequest(id, new[] { "1.0.0" }));
-
-                // Assert
-                AssertErrorResponse(
-                    controller,
-                    result,
-                    HttpStatusCode.NotFound,
-                    string.Format(Strings.DeprecatePackage_MissingVersion, id));
-                featureFlagService.Verify();
-                packageService.Verify();
-            }
-
-            [Theory]
-            [MemberData(nameof(Owner_Data))]
-            public async Task ReturnsBadRequestIfAlternateOfSelfById(User currentUser, User owner)
-            {
-                // Arrange
-                var id = "id";
-
-                var registration = new PackageRegistration
-                {
-                    Id = id
-                };
-
-                var featureFlagService = GetMock<IFeatureFlagService>();
-                featureFlagService
-                    .Setup(x => x.IsManageDeprecationEnabled(currentUser, registration))
-                    .Returns(true)
-                    .Verifiable();
-
-                registration.Owners.Add(owner);
-
-                var version = "1.0.0";
-                var package = new Package
-                {
-                    NormalizedVersion = version,
-                    PackageRegistration = registration
-                };
-
-                var packageService = GetMock<IPackageService>();
-                packageService
-                    .Setup(x => x.FindPackagesById(id, PackageDeprecationFieldsToInclude.DeprecationAndRelationships))
-                    .Returns(new[] { package })
-                    .Verifiable();
-
-                packageService
-                    .Setup(x => x.FindPackageRegistrationById(id))
-                    .Returns(registration)
-                    .Verifiable();
-
-                var controller = GetController<ManageDeprecationJsonApiController>();
-                controller.SetCurrentUser(currentUser);
-
-                // Act
-                var result = await controller.Deprecate(
-                    CreateDeprecatePackageRequest(
-                        id,
-                        new[] { version },
-                        alternatePackageId: id));
-
-                // Assert
-                AssertErrorResponse(
-                    controller,
-                    result,
-                    HttpStatusCode.BadRequest,
-                    Strings.DeprecatePackage_AlternateOfSelf);
-                featureFlagService.Verify();
-                packageService.Verify();
-            }
-
-            [Theory]
-            [MemberData(nameof(Owner_Data))]
-            public async Task ReturnsBadRequestIfAlternateOfSelfByVersion(User currentUser, User owner)
-            {
-                // Arrange
-                var id = "id";
-
-                var registration = new PackageRegistration
-                {
-                    Id = id
-                };
-
-                var featureFlagService = GetMock<IFeatureFlagService>();
-                featureFlagService
-                    .Setup(x => x.IsManageDeprecationEnabled(currentUser, registration))
-                    .Returns(true)
-                    .Verifiable();
-
-                registration.Owners.Add(owner);
-
-                var version = "1.0.0";
-                var package = new Package
-                {
-                    NormalizedVersion = version,
-                    PackageRegistration = registration
-                };
-
-                var packageService = GetMock<IPackageService>();
-                packageService
-                    .Setup(x => x.FindPackagesById(id, PackageDeprecationFieldsToInclude.DeprecationAndRelationships))
-                    .Returns(new[] { package })
-                    .Verifiable();
-
-                packageService
-                    .Setup(x => x.FindPackageByIdAndVersionStrict(id, version))
-                    .Returns(package)
-                    .Verifiable();
-
-                var controller = GetController<ManageDeprecationJsonApiController>();
-                controller.SetCurrentUser(currentUser);
-
-                // Act
-                var result = await controller.Deprecate(
-                    CreateDeprecatePackageRequest(
-                        id, 
-                        new[] { version }, 
-                        alternatePackageId: id, 
-                        alternatePackageVersion: version));
-
-                // Assert
-                AssertErrorResponse(
-                    controller,
-                    result,
-                    HttpStatusCode.BadRequest,
-                    Strings.DeprecatePackage_AlternateOfSelf);
-                featureFlagService.Verify();
-                packageService.Verify();
+                    AssertErrorResponse(controller, result, errorStatus, errorMessage);
+                }
             }
 
             private static void AssertErrorResponse(
@@ -748,227 +125,6 @@ namespace NuGetGallery.Controllers
                 Assert.Equal(error, jObject["error"].Value<string>());
             }
 
-            public static IEnumerable<object[]> PackageDeprecationStates_Data
-            {
-                get
-                {
-                    yield return MemberDataHelper.AsData(false, false, false,
-                        PackageDeprecationStatus.NotDeprecated);
-                    yield return MemberDataHelper.AsData(false, false, true,
-                        PackageDeprecationStatus.Other);
-                    yield return MemberDataHelper.AsData(false, true, false,
-                        PackageDeprecationStatus.CriticalBugs);
-                    yield return MemberDataHelper.AsData(true, false, false,
-                        PackageDeprecationStatus.Legacy);
-                    yield return MemberDataHelper.AsData(true, false, true,
-                        PackageDeprecationStatus.Legacy | PackageDeprecationStatus.Other);
-                    yield return MemberDataHelper.AsData(true, true, false,
-                        PackageDeprecationStatus.Legacy | PackageDeprecationStatus.CriticalBugs);
-                    yield return MemberDataHelper.AsData(false, true, true,
-                        PackageDeprecationStatus.CriticalBugs | PackageDeprecationStatus.Other);
-                    yield return MemberDataHelper.AsData(true, true, true,
-                        PackageDeprecationStatus.Legacy | PackageDeprecationStatus.CriticalBugs | PackageDeprecationStatus.Other);
-                }
-            }
-
-            public enum ReturnsSuccessful_AlternatePackage_State
-            {
-                None,
-                RegistrationSameId,
-                RegistrationDifferentId,
-                PackageSameId,
-                PackageDifferentId
-            }
-
-            public static IEnumerable<object[]> ReturnsSuccessful_Data =
-                MemberDataHelper.Combine(
-                    Owner_Data,
-                    PackageDeprecationStates_Data,
-                    MemberDataHelper.EnumDataSet<ReturnsSuccessful_AlternatePackage_State>()).ToList();
-
-            [Theory]
-            [MemberData(nameof(ReturnsSuccessful_Data))]
-            public Task ReturnsSuccessfulWithCustomMessage(
-                User currentUser,
-                User owner,
-                bool isLegacy,
-                bool hasCriticalBugs,
-                bool isOther,
-                PackageDeprecationStatus expectedStatus,
-                ReturnsSuccessful_AlternatePackage_State alternatePackageState)
-            {
-                return AssertSuccessful(
-                    currentUser,
-                    owner,
-                    isLegacy,
-                    hasCriticalBugs,
-                    isOther,
-                    expectedStatus,
-                    alternatePackageState,
-                    true);
-            }
-
-            /// <remarks>
-            /// Deprecations where the only reason is "other" must have a custom message.
-            /// </remarks>
-            public static IEnumerable<object[]> ReturnsSuccessfulWithoutCustomMessage_Data =
-                ReturnsSuccessful_Data.Where(x => !(bool)x[4]);
-
-            [Theory]
-            [MemberData(nameof(ReturnsSuccessfulWithoutCustomMessage_Data))]
-            public Task ReturnsSuccessfulWithoutCustomMessage(
-                User currentUser,
-                User owner,
-                bool isLegacy,
-                bool hasCriticalBugs,
-                bool isOther,
-                PackageDeprecationStatus expectedStatus,
-                ReturnsSuccessful_AlternatePackage_State alternatePackageState)
-            {
-                return AssertSuccessful(
-                    currentUser,
-                    owner,
-                    isLegacy,
-                    hasCriticalBugs,
-                    isOther,
-                    expectedStatus,
-                    alternatePackageState,
-                    false);
-            }
-
-            private async Task AssertSuccessful(
-                User currentUser,
-                User owner,
-                bool isLegacy,
-                bool hasCriticalBugs,
-                bool isOther,
-                PackageDeprecationStatus expectedStatus,
-                ReturnsSuccessful_AlternatePackage_State alternatePackageState,
-                bool hasCustomMessage)
-            {
-                // Arrange
-                var id = "id";
-
-                var registration = new PackageRegistration
-                {
-                    Id = id
-                };
-
-                var featureFlagService = GetMock<IFeatureFlagService>();
-                featureFlagService
-                    .Setup(x => x.IsManageDeprecationEnabled(currentUser, registration))
-                    .Returns(true)
-                    .Verifiable();
-
-                registration.Owners.Add(owner);
-
-                var package = new Package
-                {
-                    NormalizedVersion = "2.3.4",
-                    PackageRegistration = registration
-                };
-
-                var package2 = new Package
-                {
-                    NormalizedVersion = "1.0.0",
-                    PackageRegistration = registration
-                };
-
-                var unselectedPackage = new Package
-                {
-                    NormalizedVersion = "1.3.2",
-                    PackageRegistration = registration
-                };
-
-                var packageService = GetMock<IPackageService>();
-                packageService
-                    .Setup(x => x.FindPackagesById(id, PackageDeprecationFieldsToInclude.DeprecationAndRelationships))
-                    .Returns(new[] { package, package2, unselectedPackage })
-                    .Verifiable();
-
-                string alternatePackageId = null;
-                string alternatePackageVersion = null;
-                PackageRegistration alternatePackageRegistration = null;
-                Package alternatePackage = null;
-                if (alternatePackageState != ReturnsSuccessful_AlternatePackage_State.None)
-                {
-                    var alternateRegistration = 
-                        alternatePackageState == ReturnsSuccessful_AlternatePackage_State.RegistrationSameId 
-                        || alternatePackageState == ReturnsSuccessful_AlternatePackage_State.PackageSameId
-                        ? registration
-                        : new PackageRegistration
-                        {
-                            Id = "altId"
-                        };
-
-                    alternatePackageId = alternateRegistration.Id;
-
-                    if (alternatePackageState == ReturnsSuccessful_AlternatePackage_State.RegistrationSameId ||
-                        alternatePackageState == ReturnsSuccessful_AlternatePackage_State.RegistrationDifferentId)
-                    {
-                        alternatePackageRegistration = alternateRegistration;
-                        packageService
-                            .Setup(x => x.FindPackageRegistrationById(alternatePackageId))
-                            .Returns(alternateRegistration)
-                            .Verifiable();
-                    }
-                    else if (alternatePackageState == ReturnsSuccessful_AlternatePackage_State.PackageSameId || 
-                        alternatePackageState == ReturnsSuccessful_AlternatePackage_State.PackageDifferentId)
-                    {
-                        alternatePackageVersion = "1.2.3-alt";
-                        alternatePackage = new Package
-                        {
-                            NormalizedVersion = alternatePackageVersion,
-                            PackageRegistration = alternateRegistration
-                        };
-
-                        packageService
-                            .Setup(x => x.FindPackageByIdAndVersionStrict(alternatePackageId, alternatePackageVersion))
-                            .Returns(alternatePackage)
-                            .Verifiable();
-                    }
-                }
-
-                var deprecationService = GetMock<IPackageDeprecationService>();
-
-                var customMessage = hasCustomMessage ? "<message>" : null;
-
-                deprecationService
-                    .Setup(x => x.UpdateDeprecation(
-                        new[] { package, package2 },
-                        expectedStatus,
-                        alternatePackageRegistration,
-                        alternatePackage,
-                        customMessage,
-                        currentUser))
-                    .Completes()
-                    .Verifiable();
-
-                var controller = GetController<ManageDeprecationJsonApiController>();
-                controller.SetCurrentUser(currentUser);
-
-                var packageNormalizedVersions = new[] { package.NormalizedVersion, package2.NormalizedVersion };
-
-                // Act
-                var result = await controller.Deprecate(
-                    CreateDeprecatePackageRequest(
-                        id,
-                        packageNormalizedVersions,
-                        isLegacy,
-                        hasCriticalBugs,
-                        isOther,
-                        alternatePackageId,
-                        alternatePackageVersion,
-                        customMessage));
-
-                // Assert
-                AssertSuccessResponse(controller);
-
-                featureFlagService.Verify();
-                packageService.Verify();
-                deprecationService.Verify();
-            }
-
             private static void AssertSuccessResponse(
                 ManageDeprecationJsonApiController controller)
             {
@@ -980,29 +136,6 @@ namespace NuGetGallery.Controllers
                 HttpStatusCode code)
             {
                 Assert.Equal((int)code, controller.Response.StatusCode);
-            }
-
-            private static DeprecatePackageRequest CreateDeprecatePackageRequest(
-                string id = null,
-                IEnumerable<string> versions = null,
-                bool isLegacy = false,
-                bool hasCriticalBugs = false,
-                bool isOther = false,
-                string alternatePackageId = null,
-                string alternatePackageVersion = null,
-                string customMessage = null)
-            {
-                return new DeprecatePackageRequest
-                {
-                    Id = id,
-                    Versions = versions,
-                    IsLegacy = isLegacy,
-                    HasCriticalBugs = hasCriticalBugs,
-                    IsOther = isOther,
-                    AlternatePackageId = alternatePackageId,
-                    AlternatePackageVersion = alternatePackageVersion,
-                    CustomMessage = customMessage
-                };
             }
         }
     }
