+++ tests/NuGetGallery.Facts/NuGetGallery.Facts.csproj
@@ -96,6 +96,7 @@
     <Compile Include="Services\AsynchronousDeleteAccountServiceFacts.cs" />
     <Compile Include="Services\FeatureFlagServiceFacts.cs" />
     <Compile Include="Services\IconUrlTemplateProcessorFacts.cs" />
+    <Compile Include="Services\PackageDeprecationManagementServiceFacts.cs" />
     <Compile Include="Services\PackageDeprecationServiceFacts.cs" />
     <Compile Include="Services\SearchSideBySideServiceFacts.cs" />
     <Compile Include="Services\PackageUpdateServiceFacts.cs" />
